update & create new api giaodichmuabanlua

+++ app/Http/Services/ClientService/GiaoDichMuaBanVatTuService.php
@@ -302,6 +302,10 @@ class GiaoDichMuaBanVatTuService
             if ($request->hasFile('img_lohang')) {
                 $img_lohang = $this->uploadImageService->store($request->img_lohang);
             }
+            if($price < 0){
+                Session::flash('error', 'Giá thua mua không được nhỏ hơn 0');
+                return false;
+            }
 
             DB::beginTransaction();
             $giaodichmuabanvattu = GiaoDichMuaBanVatTu::create([
@@ -312,6 +316,7 @@ class GiaoDichMuaBanVatTuService
                 'id_category_vattu' => $id_category_vattu,
                 'img_lohang' => $img_lohang,
                 'soluong' => $soluong,
+                'price' => $price,
                 'status' => $status,
                 'description_giaodich' => $description_giaodich,
                 'hoptacxa_xacnhan' => $hoptacxa_xacnhan,
@@ -415,10 +420,18 @@ class GiaoDichMuaBanVatTuService
                 return false;
             }
 
+            if( $request->price < 0){
+                Session::flash('error', 'Giá thua mua không được nhỏ hơn 0');
+                return false;
+            }
+            
             $giaodichmuabanvattu->soluong = $request->soluong;
             $giaodichmuabanvattu->price = $request->price;
+            
+            
             $giaodichmuabanvattu->description_giaodich = $request->description_giaodich;
             $img_lohang = null;
+
             if ($request->hasFile('img_lohang')) {
                 if ($giaodichmuabanvattu->img_lohang != null) {
                     $this->uploadImageService->delete($giaodichmuabanvattu->img_lohang);
